requirements.txt
torch
torchvision
transformers
gradio
Pillow
nltk
rouge-score
rouge
pycocoevalcap
tqdm


